add debug function for verbose

+++ src/headers/assembler.h
@@ -7,4 +7,4 @@
 
 int getOpcode (std::string instr, std::bitset<4>& op, std::bitset<3>& f);
 
-int assemble (std::string line);
+unsigned short int assemble (std::string line);
